refactor: tests local fungibles
