+++ pop-api/Cargo.toml
@@ -9,9 +9,9 @@ edition = "2021"
 enumflags2 = { version = "0.7.7" }
 ink = { version = "5.0.0", default-features = false }
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
-scale-info = { version = "2.6", default-features = false, features = ["derive"] }
-sp-io = { version = "23.0.0", default-features = false, features = ["disable_panic_handler", "disable_oom", "disable_allocator"] }
-sp-runtime = { version = "24.0", default-features = false }
+scale-info = { version = "2.10.0", default-features = false, features = ["derive"] }
+sp-io = { version = "31.0.0", default-features = false, features = ["disable_panic_handler", "disable_oom", "disable_allocator"] }
+sp-runtime = { version = "32.0.0", default-features = false }
 
 pop-primitives = { path = "../primitives", features = ["devnet"], default-features = false }
