+++ pop-api/examples/.gitignore
@@ -1,9 +1,9 @@
 # Ignore build artifacts from the local tests sub-crate.
-/fungibles/target/
+**/target/
 
 # Ignore backup files creates by cargo fmt.
 **/*.rs.bk
 
 # Remove Cargo.lock when creating an executable, leave it for libraries
 # More information here http://doc.crates.io/guide.html#cargotoml-vs-cargolock
-Cargo.lock
+**/Cargo.lock
