+++ primitives/Cargo.toml
@@ -7,10 +7,10 @@ edition = "2021"
 [dependencies]
 bounded-collections = { version = "0.1", default-features = false }
 
-scale = { package = "parity-scale-codec", version = "3.6.9", default-features = false, features = ["derive"] }
+scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-decode = { version = "0.10.0", default-features = false, features = ["derive"], optional = true }
 scale-encode = { version = "0.5.0", default-features = false, features = ["derive"], optional = true }
-scale-info = { version = "2.6", default-features = false, features = ["derive"], optional = true }
+scale-info = { version = "2.10", default-features = false, features = ["derive"], optional = true }
 
 [features]
 default = ["std"]
