+++ runtime/devnet/Cargo.toml
@@ -91,7 +91,6 @@ parachain-info.workspace = true
 env_logger = "0.11.2"
 hex = "0.4.3"
 enumflags2 = "0.7.9"
-pop-api = { path = "../../pop-api", features = ["assets", "cross-chain", "nfts"] }
 
 [features]
 default = ["std"]
@@ -139,7 +138,6 @@ std = [
     "polkadot-parachain-primitives/std",
     "polkadot-runtime-common/std",
     "pop-primitives/std",
-    "pop-api/std",
     "scale-info/std",
     "sp-api/std",
     "sp-io/std",
